+++ src/diffobj.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2016  Brodie Gaslam
+ * Copyright (C) 2017  Brodie Gaslam
  *
  * This file is part of "diffobj - Diffs for R Objects"
  *