+++ tests/testthat/testthat.guide.R
@@ -9,14 +9,69 @@ rdsf <- function(x)
 test_that("detect_2d_guides", {
    iris.dply <- c("Source: local data frame [150 x 5]", "Groups: Species [3]", "", "   Sepal.Length Sepal.Width", "          (dbl)       (dbl)", "1           5.1         3.5", "2           4.9         3.0", "3           4.7         3.2", "4           4.6         3.1", "5           5.0         3.6", "6           5.4         3.9", "7           4.6         3.4", "8           5.0         3.4", "9           4.4         2.9", "10          4.9         3.1", "..          ...         ...", "Variables not shown: Petal.Length", "  (dbl), Petal.Width (dbl), Species", "  (fctr)")
    expect_equal(diffobj:::detect_2d_guides(iris.dply), 4:5)
-   old.opt <- options(width=40)
-   on.exit(options(old.opt))
-   expect_equal(diffobj:::detect_2d_guides(capture.output(iris)), c(1, 152))
-   expect_equal(
-     diffobj:::detect_2d_guides(capture.output(USAccDeaths)), c(1, 8, 15)
+   # wrapping data table with separator (#96)
+
+   DT.txt <- c(
+     "             V1        V2        V3",
+     "   1: 0.3201122 0.6907066 0.5004968",
+     "  ---                              ",
+     "1000: 0.3547379 0.2836985 0.8121208",
+     "            V4        V5",
+     "   1: 0.331665 0.6788726",
+     "  ---                   ",
+     "1000: 0.553012 0.7789110"
    )
-   # Time series
-   expect_equal(diffobj:::detect_2d_guides(capture.output(UKgas)), 1)
+   expect_equal(
+     diffobj:::detect_2d_guides(DT.txt),
+     c(1L, 5L)
+   )
+   # data table, but only if available
+
+   DT1 <- try(data.table::data.table(a=1:10), silent=TRUE)
+   if(!inherits(DT1, "try-error")) {
+     DT2 <- data.table::copy(DT1)
+     DT2[5, a:=99]
+     expect_equal(
+       as.character(diffPrint(DT1, DT2, context=1)),
+       structure(
+         c(
+           "\033[33m<\033[39m \033[33mDT1\033[39m        \033[34m>\033[39m \033[34mDT2\033[39m      ", "\033[36m@@ 5,3 @@  \033[39m  \033[36m@@ 5,3 @@  \033[39m", "\033[90m\033[90m~\033[90m \033[90m\033[90m     a\033[90m\033[90m   \033[39m  \033[90m\033[90m~\033[90m \033[90m\033[90m     a\033[90m\033[90m   \033[39m", "  \033[90m 4: \033[39m 4\033[90m\033[39m       \033[90m 4: \033[39m 4\033[90m\033[39m   ", "\033[33m<\033[39m \033[90m 5: \033[39m \033[33m5\033[39m\033[90m\033[39m     \033[34m>\033[39m \033[90m 5: \033[39m\033[34m99\033[39m\033[90m\033[39m   ",
+"  \033[90m 6: \033[39m 6\033[90m\033[39m       \033[90m 6: \033[39m 6\033[90m\033[39m   "
+         ),
+         len = 6L
+   ) ) }
+   # tibble, but only if available
+
+   TB1 <- try(tibble::tibble(a=1:10), silent=TRUE)
+   if(!inherits(TB1, "try-error")) {
+     TB2 <- TB1
+     TB2[5,"a"] <- 99L
+
+     expect_equal(
+       as.character(diffPrint(TB1, TB2, context=1)),
+       structure(
+         c(
+           "\033[33m<\033[39m \033[33mTB1\033[39m        \033[34m>\033[39m \033[34mTB2\033[39m      ",
+           "\033[36m@@ 7,3 @@  \033[39m  \033[36m@@ 7,3 @@  \033[39m",
+           "\033[90m\033[90m~\033[90m \033[90m\033[90m       a\033[90m\033[90m \033[39m  \033[90m\033[90m~\033[90m \033[90m\033[90m       a\033[90m\033[90m \033[39m",
+           "\033[90m\033[90m~\033[90m \033[90m\033[90m   <int>\033[90m\033[90m \033[39m  \033[90m\033[90m~\033[90m \033[90m\033[90m   <int>\033[90m\033[90m \033[39m",
+           "  \033[90m 4 \033[39m    4\033[90m\033[39m     \033[90m 4 \033[39m    4\033[90m\033[39m ",
+           "\033[33m<\033[39m \033[90m 5 \033[39m    \033[33m5\033[39m\033[90m\033[39m   \033[34m>\033[39m \033[90m 5 \033[39m   \033[34m99\033[39m\033[90m\033[39m ",
+           "  \033[90m 6 \033[39m    6\033[90m\033[39m     \033[90m 6 \033[39m    6\033[90m\033[39m "
+        ), len = 7L
+      )
+    )
+  }
+  # Narrow width
+
+  old.opt <- options(width=40)
+  on.exit(options(old.opt))
+  expect_equal(diffobj:::detect_2d_guides(capture.output(iris)), c(1, 152))
+  expect_equal(
+    diffobj:::detect_2d_guides(capture.output(USAccDeaths)), c(1, 8, 15)
+  )
+  # Time series
+  expect_equal(diffobj:::detect_2d_guides(capture.output(UKgas)), 1)
 })
 test_that("detect_list_guides", {
   l.1 <- list(1, 1:3, matrix(1:3, 1))